+++ src/index.ts
@@ -30,3 +30,4 @@
 export * from "./result/index.ts";
 export * from "./option/index.ts";
 export * from "./match/index.ts";
+export type { Immutable } from "./immutable/index.ts";